Simplify UI for devices that only have a power state
Only add the TelevisionSpeaker service if we have a device with mute state implemented (which is required for TelevisionSpeaker), which simplifies the UI displayed in HomeKit for things that only have a power state.

Additionally, the empty array checks in the Homebridge callbacks weren't working as it turns out `[]` is truthy - which was causing Homebridge to complain that the callback was never called. This check shouldn't actually be needed any more with the first change, but I fixed that before realising the service could be removed completely.